Refactor filtering/form state into a custom hook
